requirements.txt
Flask==2.2.2
keras==2.9.0
tensorflow-cpu==2.9.1
gunicorn
pillow